Add OpenSuSE images for Qt 5.12 and 5.14
Add a couple of images with Qt 5.12 and 5.14.
Unlike ppa:beineri, it seems there's EGL support here.

We use a different approach here: we make a huge image with all the
dependencies that we ever plan to use in our projects, to make the
workflow easier to maintain.

We build KF5 from sources, at least those few modules that we need.